% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ltrmst2adapt.R
\name{print.ltrmst2adapt}
\alias{print.ltrmst2adapt}
\title{print.ltrmst2adapt}
\usage{
\method{print}{ltrmst2adapt}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{digits}{Integer indicating the number of decimal places}

\item{...}{Further arguments ignored in this function}
}
\value{
returns summary output for class 'ltrmst2adapt'
}
\description{
S3 method for class 'ltrmst2adapt'
}
