% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rmst2adapt.R
\name{print.rmst2adapt}
\alias{print.rmst2adapt}
\title{print.rmst2adapt}
\usage{
\method{print}{rmst2adapt}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{digits}{Integer indicating the number of decimal places}

\item{...}{Further arguments ignored in this function}
}
\description{
S3 method for class 'rmst2adapt'
}
